requirements.txt
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.2
lxml>=4.9.0
Pillow>=10.0.0
win10toast>=0.9